+++ NEWS.md
@@ -6,6 +6,7 @@
 * 56 palettes have been added from the MetBrewer package. (#119)
 * 7 palettes have been added from the musculusColors package. (#120)
 * 21 palettes have been added from the peRReo package. (#122)
+* 30 palettes have been added from the NatParksPalettes package. (#123)
 
 # paletteer 1.4.1